Fix keyboard interactions jumping between values
In the port from AngularJS, we ended up conflating two `value` variables
that served different purposes. One is the actual numeric value of the
widget, and one is an internal value representing the position of the
slider around the arc of the wheel.

There was also an issue when using the keyboard to decrement the slider
value due to rounding where it would jump down by 2 numbers (but
visually put the slider in the correct(ish) spot on the wheel). That is
now also resolved.